+++ client/src/redux/session/sessionActionType.js
@@ -1,4 +1,14 @@
+export const GET_JWT_LOCAL_STORAGE = "GET_JWT_LOCAL_STORAGE"
+export const SET_JWT_LOCAL_STORAGE = "SET_JWT_LOCAL_STORAGE"
+export const CLEAR_JWT_LOCAL_STORAGE = "CLEAR_JWT_LOCAL_STORAGE"
+
+export const GET_JWT = "GET_JWT"
+export const SET_JWT = "SET_JWT"
+export const CLEAR_JWT = "CLEAR_JWT"
+
 export const FETCH_SESSION_START = 'FETCH_SESSION_START'
 export const FETCH_SESSION_SUCCESS = 'FETCH_SESSION_SUCCESS'
 export const FETCH_SESSION_FAILURE = 'FETCH_SESSION_FAILURE'
 export const FETCH_SESSION = 'FETCH_SESSION'
+
+export const KEEP_SESSION = "KEEP_SESSION"
