+++ client/src/saga/rootSaga.js
@@ -1,15 +1,11 @@
 import { all } from 'redux-saga/effects'
 
-import { watchJwt } from './authSaga'
-import { watchFetchSession } from './sessionSaga'
-import { watchKeepSession } from "./authSessionSaga"
+import { watchSession } from './sessionSaga'
 import { watchLogin } from "./loginSaga"
 
 export default function* rootSaga(){
     yield all([
-        watchJwt(),
-        watchFetchSession(),
-        watchKeepSession(),
+        watchSession(),
         watchLogin(),
     ])
 }
